Moved logout button and logic to component
