+++ frontend/src/app/[locale]/oversikt/page.tsx
@@ -15,9 +15,8 @@ import { MobileSearchBody } from '@layouts/mobile-layout/mobile-search-body.comp
 import { MobileErrandsList } from '@layouts/mobile-layout/mobile-errands-list.component';
 import { IErrand } from '@interfaces/errand';
 import { MobileMainPageHeader } from '@layouts/mobile-layout/mobile-main-page-header.component';
-import { Button, Divider, useThemeQueries } from '@sk-web-gui/react';
-import LucideIcon from '@sk-web-gui/lucide-icon';
-import { useRouter } from 'next/navigation';
+import { LogoutButton } from '@components/logout-button.component';
+import { Divider, useThemeQueries } from '@sk-web-gui/react';
 
 const Oversikt: React.FC = () => {
   const { setMunicipalityId, setUser } = useContext(AppContext);
@@ -49,11 +48,6 @@ const Oversikt: React.FC = () => {
     setOpenNotification(false);
     setOpenSearch(false);
     setOpenMenu(false);
-  };
-
-  const router = useRouter();
-  const handleLogout = () => {
-    router.push('/logout');
   };
 
   return isMaxLargeDevice ?
@@ -101,19 +95,7 @@ const Oversikt: React.FC = () => {
               <div className="py-24">
                 <Divider />
               </div>
-
-              <div className="flex justify-center w-full">
-                <Button
-                  onClick={handleLogout}
-                  variant="ghost"
-                  size="md"
-                  color="primary"
-                  className="justify-start w-full hover:bg-dark-ghost"
-                  leftIcon={<LucideIcon name="log-out" />}
-                >
-                  <span className="w-full flex justify-between">Logga ut</span>
-                </Button>
-              </div>
+              <LogoutButton />
             </MobileMenuBody>
           </MobilePage>
         )}
